refactor(podman): extract WinMemoryCheck class to dedicated file
